backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.29.0

SQLAlchemy==2.0.36
alembic==1.14.0
asyncpg==0.29.0
psycopg[binary]==3.2.3
greenlet==3.0.3

pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1

httpx==0.27.2
beautifulsoup4==4.12.3
lxml==5.3.0
tenacity==9.0.0
tqdm==4.66.4

PyJWT==2.9.0
cryptography==43.0.3

# Dev tooling
pylint==3.0.3



